+++ packages/iconoir-react-native/src/ArrowUp.tsx
@@ -14,15 +14,15 @@ function SvgArrowUp(
     <Svg
       width="1.5em"
       height="1.5em"
-      strokeWidth={1.5}
       viewBox="0 0 24 24"
+      strokeWidth={1.5}
       fill="none"
       color="currentColor"
       ref={svgRef}
       {...props}
     >
       <Path
-        d="M12.25 18.5V6m0 0l6 6m-6-6l-6 6"
+        d="M12 21V3m0 0l8.5 8.5M12 3l-8.5 8.5"
         stroke="currentColor"
         strokeLinecap="round"
         strokeLinejoin="round"