+++ packages/iconoir-react-native/src/Contactless.tsx
@@ -1,0 +1,34 @@
+import * as React from "react";
+import Svg, { SvgProps, Path } from "react-native-svg";
+import { IconoirContext } from "./IconoirContext";
+function SvgContactless(
+  passedProps: SvgProps,
+  svgRef?: React.Ref<React.Component<SvgProps>>
+) {
+  const context = React.useContext(IconoirContext);
+  const props = {
+    ...context,
+    ...passedProps,
+  };
+  return (
+    <Svg
+      width="1.5em"
+      height="1.5em"
+      viewBox="0 0 24 24"
+      strokeWidth={1.5}
+      fill="none"
+      color="currentColor"
+      ref={svgRef}
+      {...props}
+    >
+      <Path
+        d="M15 21.5c4-5.5 4-13.5 0-19M11.5 20c3.5-5 3.5-11 0-16M8.5 18c2.667-3.75 2.667-8.25 0-12M5.5 16c1.5-2.5 1.5-5.5 0-8"
+        stroke="currentColor"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </Svg>
+  );
+}
+const ForwardRef = React.forwardRef(SvgContactless);
+export default ForwardRef;