+++ packages/iconoir-react-native/src/DogecoinCircle.tsx
@@ -1,0 +1,38 @@
+import * as React from "react";
+import Svg, { SvgProps, Path } from "react-native-svg";
+import { IconoirContext } from "./IconoirContext";
+function SvgDogecoinCircle(
+  passedProps: SvgProps,
+  svgRef?: React.Ref<React.Component<SvgProps>>
+) {
+  const context = React.useContext(IconoirContext);
+  const props = {
+    ...context,
+    ...passedProps,
+  };
+  return (
+    <Svg
+      width="1.5em"
+      height="1.5em"
+      viewBox="0 0 24 24"
+      strokeWidth={1.5}
+      fill="none"
+      color="currentColor"
+      ref={svgRef}
+      {...props}
+    >
+      <Path
+        d="M10 16.402V7.598c0-.331.268-.599.6-.604 2.49-.035 5.9-.072 5.9 5.006s-3.41 5.042-5.9 5.006a.606.606 0 01-.6-.604z"
+        stroke="currentColor"
+      />
+      <Path
+        d="M8 12h4M12 22C6.477 22 2 17.523 2 12S6.477 2 12 2s10 4.477 10 10-4.477 10-10 10z"
+        stroke="currentColor"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </Svg>
+  );
+}
+const ForwardRef = React.forwardRef(SvgDogecoinCircle);
+export default ForwardRef;