+++ packages/iconoir-react/src/ArrowDown.tsx
@@ -13,8 +13,8 @@ function SvgArrowDown(
     <svg
       width="1.5em"
       height="1.5em"
-      strokeWidth={1.5}
       viewBox="0 0 24 24"
+      strokeWidth={1.5}
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
       color="currentColor"
@@ -22,7 +22,7 @@ function SvgArrowDown(
       {...props}
     >
       <path
-        d="M12.25 5.5V18m0 0l-6-6m6 6l6-6"
+        d="M12 3v18m0 0l8.5-8.5M12 21l-8.5-8.5"
         stroke="currentColor"
         strokeLinecap="round"
         strokeLinejoin="round"