+++ packages/iconoir-react/src/server/HandContactless.tsx
@@ -1,0 +1,38 @@
+import * as React from "react";
+function SvgHandContactless(
+  props: React.SVGProps<SVGSVGElement>,
+  svgRef?: React.Ref<SVGSVGElement>
+) {
+  return (
+    <svg
+      width="1.5em"
+      height="1.5em"
+      viewBox="0 0 24 24"
+      strokeWidth={1.5}
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      color="currentColor"
+      ref={svgRef}
+      {...props}
+    >
+      <path
+        d="M2 11l2.807-3.157A4 4 0 017.797 6.5H8M2 19.5h5.5l4-3s.81-.547 2-1.5c2.5-2 0-5.166-2.5-3.5C8.964 12.857 7 14 7 14"
+        stroke="currentColor"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+      <path
+        d="M8 13.5V7a2 2 0 012-2h10a2 2 0 012 2v6a2 2 0 01-2 2h-6.5"
+        stroke="currentColor"
+      />
+      <path
+        d="M18.25 12c.5-1.5.5-2.5 0-4M16 9c.227.5.227 1.5 0 2"
+        stroke="currentColor"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+}
+const ForwardRef = React.forwardRef(SvgHandContactless);
+export default ForwardRef;