+++ packages/iconoir-react/src/server/MoreHoriz.tsx
@@ -7,8 +7,8 @@ function SvgMoreHoriz(
     <svg
       width="1.5em"
       height="1.5em"
-      strokeWidth={1.5}
       viewBox="0 0 24 24"
+      strokeWidth={1.5}
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
       color="currentColor"
@@ -16,7 +16,7 @@ function SvgMoreHoriz(
       {...props}
     >
       <path
-        d="M18 12.5a.5.5 0 100-1 .5.5 0 000 1zM12 12.5a.5.5 0 100-1 .5.5 0 000 1zM6 12.5a.5.5 0 100-1 .5.5 0 000 1z"
+        d="M20 12.5a.5.5 0 100-1 .5.5 0 000 1zM12 12.5a.5.5 0 100-1 .5.5 0 000 1zM4 12.5a.5.5 0 100-1 .5.5 0 000 1z"
         fill="currentColor"
         stroke="currentColor"
         strokeLinecap="round"